backend/requirements.txt
unstructured
tiktoken
pinecone-client
pypdf
openai
langchain
pandas
numpy
python-dotenv
langchain_pinecone
django
pandas
numpy
python-dotenv
langchain
langchain-community
langchain-pinecone
langchain-cohere
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
gunicorn